fix: extract Codex hook ownership helpers under U-16 limit
Move ownership and path-resolution helpers out of codex_hooks.rs so the
file stays under the 800-line hard limit and restore Core CI.
