+++ vibeguard-runtime/src/setup/codex_hooks/tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use std::path::PathBuf;
 
 fn repo_dir() -> PathBuf {
     Path::new(env!("CARGO_MANIFEST_DIR")).join("..")
